Add OpenAPI spec and serve it using SwaggerUI
Create `fiberswagger` middleware to serve SwaggerUI. Although there is
a swagger middleware available on the `fiber/contrib` repository that
implementation if not compatiblewith OpenAPI 3.0.x specs, so in order
to support the new standard this middleware was built to be compatible
with Swagger 2.0 specs and OpenAPI 3.0.x specs.

Also fix some error messages and conflicting LICENSE definition.

Signed-off-by: Filipe Utzig <filipeutzig@gmail.com>