Create basic debian package
Setup basic infrastructure and metadata for building a simple debian
package to install the project.

Known limitations includes building support only for the amd64 arch.

The goal of the project is to build debian deployable packages for ARM
architectures as well, but this should be achieved in upcoming commits.

Docker image was created to setup all dependencies for generating the
deb package and a handy script is shipped with the image to call the
debian scripts properly and copy the build artifacts into the project
source tree. Also a basic Makefile with targets for building the deb
package was created to help daily usage.

Signed-off-by: Filipe Utzig <filipeutzig@gmail.com>