Fix fiberswagger middleware routing issues
The fiberswagger middleware wasn't calling the next handler after
proccesing the request and since it is mounted on the root router no
requests other than the ones for itself were being served, causing all
other routes to fail with 404 error.

Fixed the issue using the adaptor.HTTPMiddleware method instead of the
adaptor.HTTPHandler to return a proper fiber.Handler.

Signed-off-by: Filipe Utzig <filipeutzig@gmail.com>